+++ src/components/forms/ExperienceForm/pages/Page3.tsx
@@ -1,114 +1,161 @@
-import {
-  TextField,
-  Autocomplete,
-  SelectChangeEvent,
-} from '@mui/material';
+import { Box, Button, styled } from '@mui/material';
 import { FormPageProps } from '../formPageProps';
-import React, { ChangeEvent, useEffect } from 'react';
-import selectOptions from "../selectOptions.json";
+import { CloudUpload as CloudUploadIcon } from '@mui/icons-material';
+import React, { useEffect, ChangeEvent } from 'react';
+import { useValidation, ValidationRule } from '../../../../hooks/useValidation';
 
-const Page3: React.FC<FormPageProps> = ({
+interface ImageUploadPageProps extends FormPageProps {
+  foodPhoto: File | null;
+  personPhoto: File | null;
+  setFoodPhoto: (photo: File) => void;
+  setPersonPhoto: (photo: File) => void;
+};
+
+const Input = styled('input')({
+  display: 'none',
+});
+
+const ImagePreview = styled('img')({
+  width: '100%', // This can be adjusted to fit your design
+  height: 'auto',
+  marginBottom: '10px',
+  border: '1px solid #ddd',
+  borderRadius: '4px',
+  padding: '5px',
+});
+
+const Page4: React.FC<ImageUploadPageProps> = ({
   experience,
   setExperience,
+  foodPhoto,
+  personPhoto,
+  setFoodPhoto,
+  setPersonPhoto,
+  setError,
   onValidationSuccess,
   validationTrigger,
   setValidationTrigger
 }) => {
-  const handleComboBoxChange = (newValue: any, fieldName: string) => {
-    const value = Array.isArray(newValue) ? newValue.join(",") : newValue;
+  const handleFileUpload = (event: ChangeEvent<HTMLInputElement>, photoType: 'food' | 'person') => {
+    setError("");
+    const file = event.target.files ? event.target.files[0] : null;
 
-    setExperience({ ...experience, [fieldName]: value});
+    if (!file) {
+      console.log("No file!");
+      return;
+    }
+
+    const MAX_IMAGE_SIZE = 1024 * 1024 * 5;
+    if (file.size > MAX_IMAGE_SIZE) {
+      setError(`Max file size exceeded. Max file size: ${MAX_IMAGE_SIZE / (1024 * 1024)} MB`);
+      return;
+    }
+
+    const allowedFileTypes = ["image/png", "image/jpeg"];
+    if (!allowedFileTypes.includes(file.type)) {
+      setError(`File type not allowed. Allowed file types: ${allowedFileTypes.join(", ")}`);
+      return;
+    }
+
+    const reader = new FileReader();
+    reader.onloadend = () => {
+      if (photoType === 'food') {
+        setFoodPhoto(file);
+        setExperience({
+          ...experience,
+          foodPhotoUrl: reader.result as string
+        });
+      } else {
+        setPersonPhoto(file);
+        setExperience({
+          ...experience,
+          personPhotoUrl: reader.result as string
+        });
+      }
+    };
+    reader.readAsDataURL(file);
   };
+
+  const validationRules = [
+    {
+      errorCondition: !foodPhoto && !experience._id,
+      message: "Please select a food photo",
+      fieldName: "foodPhoto"
+    }
+  ] as ValidationRule[];
+
+  const { validateFields } = useValidation(
+    ["foodPhoto", "personPhoto"],
+    validationRules,
+    setError,
+    onValidationSuccess
+  );
 
   useEffect(() => {
     if (validationTrigger) {
+      validateFields();
       setValidationTrigger(false);
-      onValidationSuccess();
     }
-  }, [validationTrigger, setValidationTrigger, onValidationSuccess]);
+  }, [validationTrigger, setValidationTrigger, validateFields])
 
   return (
     <>
-      <Autocomplete
-        onChange={(event, newValue) => handleComboBoxChange(newValue, "personItRemindsThemOf")}
-        data-testid="ExperienceForm-PersonItRemindsThemOfField"
-        value={experience.personItRemindsThemOf?.split(",")}
-        options={selectOptions.relationships}
-        freeSolo
-        multiple
-        fullWidth
-        renderInput={(params) => <TextField
-          {...params}
-          label="Person/people it reminds you of (relationship)"
-        />}
-      />  
-      <Autocomplete
-        onChange={(event, newValue) => handleComboBoxChange(newValue, "periodOfLifeAssociation")}
-        data-testid="ExperienceForm-PeriodOfLifeField"
-        value={experience.periodOfLifeAssociation?.split(",")}
-        options={selectOptions.periodsOfLife}
-        freeSolo
-        multiple
-        fullWidth
-        renderInput={(params) => <TextField
-          {...params}
-          label="Period of Life Association"
-        />}
-      />   
-      <Autocomplete
-        onChange={(event, newValue) => handleComboBoxChange(newValue, "mood")}
-        data-testid="ExperienceForm-EmotionsField"
-        value={experience.mood?.split(",")}
-        options={selectOptions.emotions}
-        freeSolo
-        multiple
-        fullWidth
-        renderInput={(params) => <TextField
-          {...params}
-          label="Emotions Felt"
-        />}
-      />  
-      <Autocomplete
-        onChange={(event, newValue) => handleComboBoxChange(newValue, "foodtype")}
-        data-testid="ExperienceForm-MealTypeField"
-        value={experience.foodtype?.split(",")}
-        options={selectOptions.mealTypes}
-        freeSolo
-        multiple
-        fullWidth
-        renderInput={(params) => <TextField
-          {...params}
-          label="Meal Type"
-        />}
-      /> 
-      <Autocomplete
-        onChange={(event, newValue) => handleComboBoxChange(newValue, "flavourProfile")}
-        data-testid="ExperienceForm-FlavourProfileField"
-        value={experience.flavourProfile?.split(",")}
-        options={selectOptions.flavorProfiles}
-        freeSolo
-        multiple
-        fullWidth
-        renderInput={(params) => <TextField
-          {...params}
-          label="Flavour Profile"
-        />}
-      />
-      <Autocomplete
-        onChange={(event, newValue) => handleComboBoxChange(newValue, "cuisine")}
-        data-testid="ExperienceForm-CuisineField"
-        value={experience.cuisine?.split(",")}
-        options={selectOptions.cuisines}
-        freeSolo
-        multiple
-        fullWidth
-        renderInput={(params) => <TextField
-          {...params}
-          label="Cuisine"
-        />}
-      /> 
+      <Box
+        sx={{
+          display: "flex",
+          justifyContent: "center",
+          alignContent: "center",
+          gap: "20px",
+        }}
+      >
+        {/* Upload and preview for food photo */}
+        <Box sx={{ display: 'flex', flexDirection: 'column', alignItems: 'center' }}>
+          {experience.foodPhotoUrl && <ImagePreview src={experience.foodPhotoUrl} alt="Food preview" data-testid="ExperienceForm-FoodPhotoPreview" />}
+          <label htmlFor="food-photo-upload">
+            <Input
+              accept="image/*"
+              id="food-photo-upload"
+              type="file"
+              onChange={(event) => handleFileUpload(event, 'food')}
+              data-testid="ExperienceForm-UploadFoodPhotoInput"
+            />
+            <Button
+              variant="contained"
+              color="primary"
+              component="span"
+              startIcon={<CloudUploadIcon />}
+              data-testid="ExperienceForm-UploadFoodPhotoButton"
+            >
+              Food Photo
+            </Button>
+          </label>
+        </Box>
+
+        {/* Upload and preview for person photo */}
+        <Box sx={{ display: 'flex', flexDirection: 'column', alignItems: 'center' }}>
+          {experience.personPhotoUrl && <ImagePreview src={experience.personPhotoUrl} alt="Person preview" data-testid="ExperienceForm-PersonPhotoPreview" />}
+          <label htmlFor="person-photo-upload">
+            <Input
+              accept="image/*"
+              id="person-photo-upload"
+              type="file"
+              onChange={(event) => handleFileUpload(event, 'person')}
+              data-testid="ExperienceForm-UploadPersonPhotoInput"
+            />
+            <Button
+              variant="contained"
+              color="primary"
+              component="span"
+              startIcon={<CloudUploadIcon />}
+              data-testid="ExperienceForm-UploadPersonPhotoButton"
+            >
+              Person Photo (Optional)
+            </Button>
+          </label>
+        </Box>
+      </Box>
     </>
   );
 };
 
-export default Page3;
+export default Page4;
