+++ src/components/forms/ExperienceForm/pages/index.ts
@@ -1,4 +1,3 @@
 export { default as Page1 } from './Page1';
 export { default as Page2 } from './Page2';
 export { default as Page3 } from './Page3';
-export { default as Page4 } from './Page4';
